# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyzeRequest {
  /**
   * Conversation fragment to analyze
   * @minLength 1
   */
  text: string;
  /** Optional session context to guide the tactical analysis */
  context?: string;
}

export interface AnalyzeResponse {
  /** Short signal label (2-5 words) */
  signal: string;
  /** Tactical action to take (4-12 words) */
  say_now: string;
  /** What to avoid doing right now (2-6 words) */
  avoid: string;
}

export interface ErrorResponse {
  error: string;
}